requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
gunicorn
python-dotenv
pymongo
dnspython